refactor: 채팅 도메인 패키지 구조 DDD 패턴으로 개선
- presentation, application, domain, infrastructure 계층 분리
- controller → presentation/controller 이동
- model → presentation/dto로 명확화
- repository 인터페이스를 application 계층으로 이동
- valide → validation으로 패키지명 개선
