+++ src/main/java/com/debateseason_backend_v1/domain/chat/domain/model/Chat.java
@@ -1,0 +1,4 @@
+package com.debateseason_backend_v1.domain.chat.domain.model;
+
+public class Chat {
+}
