+++ src/main/java/com/debateseason_backend_v1/domain/chat/presentation/dto/response/ChatMessageErrorResponse.java
